Make first migration

+++ infrastructure/database/models/profit.py
@@ -1,5 +1,4 @@
-from sqlalchemy import Column, Text, Integer, ForeignKey
-from sqlalchemy import Column, Text, Integer, ForeignKey
+from sqlalchemy import Column, Text, Integer, ForeignKey, BigInteger
 from sqlalchemy.orm import relationship
 
 from infrastructure.database.models.base import Base, TimestampMixin, TableNameMixin
@@ -7,7 +6,7 @@ from infrastructure.database.models.base import Base, TimestampMixin, TableNameM
 
 class Profit(Base, TimestampMixin, TableNameMixin):
     profit_id = Column(Integer, primary_key=True)
-    owner_id = Column(Integer, ForeignKey("users.user_id"), nullable=False)
+    owner_id = Column(BigInteger, ForeignKey("users.user_id"), nullable=False)
     amount = Column(Integer, nullable=False)
     raw_text = Column(Text)
